Added a basic circle and applying rendering hints

+++ src/com/Game/Objects/Entity.java
@@ -2,5 +2,5 @@ package com.Game.Objects;
 
 public interface Entity {
     
-    public abstract void update(double dt);
+    public abstract void update(float dt);
 }
